00_rendering/AtmosphereRenderer/requirements.txt
numpy
scipy
matplotlib
tqdm
Pillow
h5py
scikit-learn
pyvista
vtk
```